chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` and `yarn.lock` files. It adds the `apicache` and `cors` packages, and updates the versions of `dotenv`, `express`, and `express-rate-limit`. These changes were made to ensure that the project is using the latest stable versions of the dependencies and to introduce new functionality for caching API responses and handling CORS.