Feat/setup rate limiting and apicache (#2)
* chore: Update npm dependencies

This commit updates the npm dependencies in the `package.json` and `yarn.lock` files. It adds the `apicache` and `cors` packages, and updates the versions of `dotenv`, `express`, and `express-rate-limit`. These changes were made to ensure that the project is using the latest stable versions of the dependencies and to introduce new functionality for caching API responses and handling CORS.

* chore: Update routes.js to include caching and rate limiting

* chore: Update TMDB and Trakt API base URLs

This commit updates the TMDB and Trakt API base URLs in the `tmdb.js` and `trakt.js` files respectively. The base URLs are now retrieved from environment variables (`process.env.TMDB_API_BASE_URL` and `process.env.TRAKT_API_BASE_URL`) instead of being hard-coded. These changes were made to improve flexibility and allow for easier configuration of the API endpoints.